# docs/openapi-daemon.yaml
components:
  securitySchemes:
    bearerToken:
      description: 'The daemon API token: SENCLAW_API_TOKEN, else the generated ~/.senclaw/api_token.'
      scheme: bearer
      type: http
    sessionCookie:
      description: Minted by POST /api/auth/login.
      in: cookie
      name: senclaw_token
      type: apiKey
    tokenHeader:
      description: The same token, for clients that cannot set Authorization.
      in: header
      name: X-SenClaw-Token
      type: apiKey
info:
  description: |-
    Generated from the daemon's own routers. Paths, methods and path parameters are exact. Bodies are not described: handlers that return `Json<serde_json::Value>` are marked `x-untyped: true`, and `x-untyped-operations` counts how many remain.

    A few templates overlap (`/api/oauth/{provider}/start` and `/api/oauth/accounts/{id}`, for example). That is not a conflict at runtime: axum matches a literal segment before a capture, so `/api/oauth/accounts/1` reaches the accounts handler. A client generated from this document should prefer the more literal template for the same reason.
  title: SenClaw daemon API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/browser-agent/approvals/{id}:
    post:
      operationId: post_api_browser_agent_approvals_id
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: Success
        '401':
          description: Token required or invalid (see SENCLAW_AUTH_MODE)
      summary: Approval
      tags:
      - browser-agent
      x-handler: approval
      x-untyped: true
  /api/browser-agent/do:
    post:
      operationId: post_api_browser_agent_do
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: Success
        '401':
          description: Token required or invalid (see SENCLAW_AUTH_MODE)
      summary: Do step
      tags:
      - browser-agent
      x-handler: do_step
      x-untyped: true
...
  /api/browser-agent/handover:
    post:
      operationId: post_api_browser_agent_handover
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: Success
        '401':
          description: Token required or invalid (see SENCLAW_AUTH_MODE)
      summary: Handover
      tags:
      - browser-agent
      x-handler: handover
      x-untyped: true
  /api/browser-agent/look:
    post:
      operationId: post_api_browser_agent_look
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: Success
        '401':
          description: Token required or invalid (see SENCLAW_AUTH_MODE)
      summary: Look
      tags:
      - browser-agent
      x-handler: look
      x-untyped: true
  /api/browser-agent/open:
    post:
      operationId: post_api_browser_agent_open
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: Success
        '401':
          description: Token required or invalid (see SENCLAW_AUTH_MODE)
      summary: Open
      tags:
      - browser-agent
      x-handler: open
      x-untyped: true
  /api/browser-agent/read:
    post:
      operationId: post_api_browser_agent_read
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: Success
        '401':
          description: Token required or invalid (see SENCLAW_AUTH_MODE)
      summary: Read
      tags:
      - browser-agent
      x-handler: read
      x-untyped: true
  /api/browser-agent/screenshot:
    post:
      operationId: post_api_browser_agent_screenshot
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: Success
        '401':
          description: Token required or invalid (see SENCLAW_AUTH_MODE)
      summary: Screenshot
      tags:
      - browser-agent
      x-handler: screenshot
      x-untyped: true
...
  /api/browser-agent/tasks:
    post:
      operationId: post_api_browser_agent_tasks
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: Success
        '401':
          description: Token required or invalid (see SENCLAW_AUTH_MODE)
      summary: Task
      tags:
      - browser-agent
      x-handler: task
      x-untyped: true
  /api/browser-agent/tasks/{id}/resume:
    post:
      operationId: post_api_browser_agent_tasks_id_resume
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: Success
        '401':
          description: Token required or invalid (see SENCLAW_AUTH_MODE)
      summary: Task resume
      tags:
      - browser-agent
      x-handler: task_resume
      x-untyped: true
  /api/browser/{rest}:
    get:
      operationId: get_api_browser_rest
      parameters:
      - in: path
        name: rest
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: Success
        '401':
          description: Token required or invalid (see SENCLAW_AUTH_MODE)
      summary: Proxy browser
      tags:
      - browser
      x-handler: proxy_browser
      x-untyped: true
...
  /api/space/apps/{id}/mcp:
    get:
      operationId: get_api_space_apps_id_mcp
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: Success
        '401':
          description: Token required or invalid (see SENCLAW_AUTH_MODE)
      summary: Space app mcp info
      tags:
      - space
      x-handler: space_app_mcp_info
      x-untyped: true
...
security:
- bearerToken: []
- tokenHeader: []
- sessionCookie: []
servers:
- description: Local daemon
  url: http://127.0.0.1:18788
...